fixed config
