+++ .gitignore
@@ -16,6 +16,7 @@ package-lock.json
 
 # misc
 .DS_Store
+.env
 .env.local
 .env.development.local
 .env.test.local
