+++ src/components/LoginModal/MainContent/MainContent.jsx
@@ -1,8 +1,5 @@
-import React ,{ useContext } from 'react';
+import React from 'react';
 import styled from 'styled-components';
-import { AuthContext } from '../../../contexts/AuthContext';
-import firebase from '../../../plugins/firebase';
-
 import { StyledLoginForm as LoginForm } from './LoginForm/LoginForm';
 import { StyledHorizon as Horizon } from './Horizon/Horizon';
 import { StyledGoogleLoginButton as GoogleLoginButton } from './GoogleLogin/GoogleLogin';
@@ -15,21 +12,11 @@ const StyledMain = styled.main`
 `;
 
 export const MainContent = () => {
-    const { auth} = useContext(AuthContext);
-    // For firebase Google Auth
-    const provider = new firebase.auth.GoogleAuthProvider();
-    const handleLogin = () => {
-        firebase.auth().signInWithPopup(provider)
-            .then( user => {
-                console.log('user from login button: ',user)
-            })
-            .catch(err => console.log(err))
-    }
     return (
-        <StyledMain onClick={handleLogin}>
+        <StyledMain>
             <LoginForm />
             <Horizon />
-            <GoogleLoginButton />
+            <GoogleLoginButton/>
         </StyledMain>
     );
 }
