--- src/plugins/createdInBranch.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const createdInBranch = () => {
-    return(
-        <h1>Hi</h1>
-    )
-}
-
-export default createdInBranch;
